updated notes routes with folder info and tests

+++ routes/notes.js
@@ -8,15 +8,19 @@ const router = express.Router();
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
-  const {searchTerm} = req.query;
+  const {searchTerm, folderId} = req.query;
+  const folderFilter = {};
+  if (folderId) {
+    folderFilter.folderId = folderId;
+  }
+
   if (searchTerm) {
     Note.find(
       {$or:
-            [
-              {title: {$regex: searchTerm, $options: 'i'}},
-              {content: {$regex: searchTerm, $options: 'i'}}
-            ]
+            [ {title: {$regex: searchTerm, $options: 'i'}},
+              {content: {$regex: searchTerm, $options: 'i'}} ]
       })
+      .find(folderFilter)
       .sort('_id')
       .then(results => {
         res.json(results);
@@ -25,7 +29,7 @@ router.get('/', (req, res, next) => {
         next(err);
       });
   } else{
-    Note.find()
+    Note.find(folderFilter)
       .sort('_id')
       .then(results => {
         res.json(results);
@@ -41,6 +45,11 @@ router.get('/', (req, res, next) => {
 router.get('/:id', (req, res, next) => {
  
   const {id} = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
   return Note.findById(id)
 
     .then(result => {
@@ -58,16 +67,26 @@ router.get('/:id', (req, res, next) => {
 
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
-  const {title, content} = req.body;
+  const {title, content, folderId} = req.body;
   const newNote = {
     title: title, 
     content: content, 
+    folderId: folderId
   };
-  if (!newNote.title || !newNote.content) {
-    const err = new Error('Missing `title` or `content` in request body');
+
+    /*validation - check if id is valid and if exists*/
+  if (!mongoose.Types.ObjectId.isValid(folderId)) {
+    const err = new Error('The `folderId` is not valid');
     err.status = 400;
     return next(err);
   }
+
+  if (!newNote.title || !newNote.folderId) {
+    const err = new Error('Missing `title` or `folderId` in the request body');
+    err.status = 400;
+    return next(err);
+  }
+
   return Note.create(newNote)
     .then(results => {
       if (results){
@@ -82,21 +101,30 @@ router.post('/', (req, res, next) => {
 /* ========== PUT/UPDATE A SINGLE ITEM ========== */
 router.put('/:id', (req, res, next) => {
   const {id} = req.params;
-  const {title, content} = req.body;
+  const {title, content, folderId} = req.body;
   const updatedNote = {
     title: title, 
     content: content, 
+    folderId: folderId
   };
-  if (!updatedNote.title) {
-    const err = new Error('Missing `title` in request body');
+
+  /*validation - check if id is valid and if exists*/
+  if (!mongoose.Types.ObjectId.isValid(id) || !mongoose.Types.ObjectId.isValid(folderId)) {
+    const err = new Error('The `id` or `folderId` is not valid');
+    err.status = 400;
+    return next(err);
+  }
+
+  if (!updatedNote.title || !updatedNote.folderId) {
+    const err = new Error('Missing `title` or `folderId` in the request body');
     err.status = 400;
     return next(err);
   }
 
   return Note.findByIdAndUpdate(id, updatedNote, {new:true})
-    .then(results => {
-      if (results){
-        res.json(results);
+    .then(result => {
+      if (result){
+        res.json(result);
       } else {
         next();
       }
@@ -108,6 +136,14 @@ router.put('/:id', (req, res, next) => {
 router.delete('/:id', (req, res, next) => {
 
   const {id} = req.params;
+
+  /***** Never trust users - validate input *****/
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  } 
+
   return Note.findByIdAndRemove(id)
     .then(()=>{
       res.status(204).end();
